requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
flask-limiter==3.5.0
gunicorn==21.2.0
